Add VTT Notes field to common row data.

+++ com.trollworks.gcs/src/com/trollworks/gcs/body/HitLocation.java
@@ -209,7 +209,7 @@ public class HitLocation implements Cloneable, Comparable<HitLocation> {
         if (!dr.containsKey(DRBonus.ALL_SPECIALIZATION)) {
             dr.put(DRBonus.ALL_SPECIALIZATION, Integer.valueOf(0));
         }
-        int          all  = dr.get(DRBonus.ALL_SPECIALIZATION);
+        int          all  = dr.get(DRBonus.ALL_SPECIALIZATION).intValue();
         List<String> keys = new ArrayList<>(dr.keySet());
         keys.remove(DRBonus.ALL_SPECIALIZATION);
         Collections.sort(keys);
